Move StartDurableExternalEventListenerCommand to Durable\Commands
